// src/types/HeaderTypes.ts
type notificationType = {
    avatar: string;
    title: string;
    subtitle: string;
};

type profileType = {
    avatar: string;
    title: string;
    subtitle: string;
    href: string;
};

type languageType = {
    avatar: string;
    title: string;
    subtext: string;
    value: string;
};


type searchType = {
    title: string;
    href: string;
};


export type { notificationType, profileType, languageType,searchType }